Add crawler package
Define a basic crawler implementation that accepts a start URL and
a number of concurrent workers to use for crawling web pages. The
crawler uses multiple workers for performing concurrent work and a
single processor for processing results retrieved by workers.

The crawler package also defines a Fetcher interface to use for
fetching links from a given target URL. We define a fake fetcher
that fulfils the interface to aid testing of our concurrent model.